美食分类维护

+++ src/main/java/com/myicellar/digitalmenu/controller/manage/FoodTypeManageController.java
@@ -1,8 +1,10 @@
 package com.myicellar.digitalmenu.controller.manage;
 
+import com.aliyuncs.utils.StringUtils;
 import com.myicellar.digitalmenu.dao.entity.FoodType;
 import com.myicellar.digitalmenu.service.FoodTypeService;
 import com.myicellar.digitalmenu.shiro.AuthIgnore;
+import com.myicellar.digitalmenu.utils.BizException;
 import com.myicellar.digitalmenu.utils.ConvertUtils;
 import com.myicellar.digitalmenu.utils.SnowflakeIdWorker;
 import com.myicellar.digitalmenu.vo.request.FoodTypeDeleteReqVO;
@@ -72,6 +74,8 @@ public class FoodTypeManageController {
     @AuthIgnore
     @ApiOperation("新增")
     public ResultVO add(@RequestBody FoodTypeReqVO reqVO) {
+        //参数校验
+        checkNewParam(reqVO);
         FoodType foodType = ConvertUtils.convert(reqVO,FoodType.class);
         foodType.setFoodTypeId(snowflakeIdWorker.nextId());
         foodType.setCreateUser(0L);
@@ -101,6 +105,8 @@ public class FoodTypeManageController {
     @AuthIgnore
     @ApiOperation("修改")
     public ResultVO update(@RequestBody FoodTypeReqVO reqVO) {
+        //参数校验
+        checkUpdateParam(reqVO);
         FoodType foodType = ConvertUtils.convert(reqVO,FoodType.class);
         foodType.setUpdateUser(0L);
         Date now = new Date();
@@ -127,12 +133,44 @@ public class FoodTypeManageController {
     @AuthIgnore
     @ApiOperation("删除")
     public ResultVO update(@RequestBody FoodTypeDeleteReqVO reqVO) {
+        if(reqVO.getFoodTypeId()==null || reqVO.getFoodTypeId()==0L){
+            return ResultVO.validError("parameter is invalid！");
+        }
         int i = foodTypeService.deleteByPrimaryKey(reqVO.getFoodTypeId());
         if(i==0){
             return ResultVO.validError("delete is failed!");
         }
 
         return ResultVO.success("delete is success!");
+    }
+
+    /**
+     * 校验新增参数
+     * @param reqVO
+     */
+    public void checkNewParam(FoodTypeReqVO reqVO){
+        if(reqVO.getSupplierId()==null || reqVO.getSupplierId()==0L){
+            throw new BizException("supplier cannot is empty!");
+        }
+        if(StringUtils.isEmpty(reqVO.getFoodTypeNameEng())){
+            throw new BizException("food_type_name_eng cannot is empty!");
+        }
+    }
+
+    /**
+     * 校验修改参数
+     * @param reqVO
+     */
+    public void checkUpdateParam(FoodTypeReqVO reqVO){
+        if(reqVO.getFoodTypeId()==null || reqVO.getFoodTypeId()==0L){
+            throw new BizException("foodTypeId cannot is empty!");
+        }
+        if(reqVO.getSupplierId()==null || reqVO.getSupplierId()==0L){
+            throw new BizException("supplier cannot is empty!");
+        }
+        if(StringUtils.isEmpty(reqVO.getFoodTypeNameEng())){
+            throw new BizException("foodTypeNameEng cannot is empty!");
+        }
     }
 
 }
